+++ tests/integ/scala/test_dataframe_join_suite.py
@@ -1667,3 +1667,30 @@ def test_dataframe_alias_map_unmodified(session):
         Utils.check_answer(aggregate(df), [Row("James", 3000)])
     finally:
         session._join_alias_fix = origin
+
+
+@pytest.mark.parametrize("with_condition", [True, False])
+@pytest.mark.skipif(
+    "config.getoption('local_testing_mode', default=False)",
+    reason="Lateral join is not supported in Local Testing",
+)
+def test_lateral_join_behaviors(session, with_condition):
+    df1 = session.create_dataframe([[1, 2], [3, 4], [5, 6]], schema=["a", "b"])
+    df2 = session.create_dataframe([[1, 7], [3, 8]], schema=["a", "c"])
+
+    condition = df1.a == df2.a if with_condition else None
+    df = df1.lateral_join(df2, condition)
+
+    # expected results
+    matching_rows = [Row(1, 2, 1, 7), Row(3, 4, 3, 8)]
+    cartesian_product = [
+        Row(1, 2, 1, 7),
+        Row(1, 2, 3, 8),
+        Row(3, 4, 1, 7),
+        Row(3, 4, 3, 8),
+        Row(5, 6, 1, 7),
+        Row(5, 6, 3, 8),
+    ]
+
+    expected = matching_rows if with_condition else cartesian_product
+    Utils.check_answer(df, expected)