+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 #### New Features
 
+- Added support for `DataFrame.lateral_join`
 - Added support for `Session.client_telemetry`.
 - Added support for `Session.udf_profiler`.
 - Added support for `functions.ai_translate`.
@@ -56,6 +57,7 @@
 
 - Enhanced `DataFrame.sort()` to support `ORDER BY ALL` when no columns are specified.
 - Catalog API now uses SQL commands instead of SnowAPI calls. This new implementation is more reliable now.
+- Removed experimental warning from `Session.cte_optimization_enabled`.
 
 #### Dependency Updates
 
@@ -134,6 +136,10 @@
   - `dt.month_name`
   - `dt.day_name`
   - `dt.strftime`  
+  - `dt.dayofweek`  
+  - `dt.weekday`  
+  - `dt.dayofyear`    
+  - `dt.isocalendar`  
   - `rolling.min`
   - `rolling.max`
   - `rolling.count`
@@ -154,6 +160,37 @@
   - `cumsum`
   - `cummin`
   - `cummax`
+  - `groupby.groups`
+  - `groupby.indices`
+  - `groupby.first`
+  - `groupby.last`
+  - `groupby.rank`
+  - `groupby.shift`
+  - `groupby.cumcount`
+  - `groupby.cumsum`
+  - `groupby.cummin`
+  - `groupby.cummax`
+  - `groupby.any`
+  - `groupby.all`
+  - `groupby.unique`
+  - `groupby.get_group`
+  - `groupby.rolling`
+  - `groupby.resample`
+  - `to_snowflake`
+  - `to_snowpark`
+  - `resample.min`
+  - `resample.max`
+  - `resample.count`
+  - `resample.sum`
+  - `resample.mean`
+  - `resample.median`
+  - `resample.std`
+  - `resample.var`
+  - `resample.size`
+  - `resample.first`
+  - `resample.last`
+  - `resample.quantile`
+  - `resample.nunique`
 - Make faster pandas disabled by default (opt-in instead of opt-out).
 - Improve performance of `drop_duplicates` by avoiding joins when `keep!=False` in faster pandas.